requirements.txt
pyTelegramBotAPI
phidata
google-genai
groq
tavily-python
exa-py
duckduckgo-search
jina
python-dotenv
